fix bug 整数相除时要强转小数

+++ app/src/main/java/com/etatech/test/utils/App.java
@@ -33,7 +33,7 @@ public class App extends Application {
 
         screenHeight = ScreenUtils.getScreenHeight();
         screenWidth = ScreenUtils.getScreenWidth();
-        if (screenHeight / screenWidth >= designRatio) {
+        if (screenHeight / (float) screenWidth >= designRatio) {
             isWide = false;
             LogUtils.i(String.format("The App Screen Height %d Width %d ", screenHeight, screenWidth));
         } else {
